+++ typedAPI/http/response/factory.py
@@ -1,11 +1,8 @@
-
-
 import starlette.responses
-
 from typedAPI.http.response.data import http_status_codes
 
 
-def make_response_from_status(status: int, content_type="application/json") -> starlette.responses.Response:
+def make_response_from_status(status: int, content_type = "application/json") -> starlette.responses.Response:
     
     message = http_status_codes[status]
 
@@ -21,4 +18,21 @@ def make_response_from_status(status: int, content_type="application/json") -> s
         body = { "error": "Unsupported content type: " + content_type }
     
     return starlette.responses.Response(body, status_code=status, media_type=content_type)
+
+
+
+def make_complete_headers_response(status, body_as_byes, content_type) -> starlette.responses.Response:
+    return starlette.responses.Response(
+        body_as_byes,
+        status_code = status,
+        headers = { "content-type": content_type }
+    )
+
+
+def make_response_from_complete_spec(status, headers, body) -> starlette.responses.Response:
+    return starlette.responses.Response(
+        body,
+        status_code = status,
+        headers = headers
+    )
