+++ typedAPI/http/response/service.py
@@ -1,12 +1,9 @@
 
 
 from typedAPI.http.response.schema import Response, Status, Body, Headers, NormalisedResponse
-from typedAPI.http.response.factory import make_response_from_status
-from typedAPI.http.response.data import guess_content_type_from_body, cast_from_content_type
-
-
+from typedAPI.http.response.factory import make_response_from_status, make_complete_headers_response, make_response_from_complete_spec
+from typedAPI.http.response.processors import guess_content_type_from_body, cast_from_content_type
 import typing
-
 import starlette.responses
 
 
@@ -20,12 +17,9 @@ def is_response(values: typing.Any) -> typing.TypeGuard[Response]:
     
     size = len(values)
 
-    if size > 3:
+    if size > 3 or size == 0:
         return False
-    
-    if size == 0:
-        return False
-    
+
     status = values[0]
     
     if not isinstance(status, Status):
@@ -60,25 +54,15 @@ def normalise_response(response: Response) -> NormalisedResponse:
     
     if size == 2:
         return response[0], response[1], ...
+    
+    return response[0], response[1], response[2] # type: ignore
 
-    return response # type: ignore
-
-
-
-
-def make_complete_headers_response(status, body_as_byes, content_type) -> starlette.responses.Response:
-    return starlette.responses.Response(
-        body_as_byes,
-        status_code = status,
-        headers = { "content-type": content_type }
-    )
 
 
 def to_starlette_response(normalised_response: NormalisedResponse) -> starlette.responses.Response:
     status, headers, body = normalised_response
     
     complete_headers = headers == ...
-
     complete_body = body == ...
 
     if complete_headers:
@@ -91,25 +75,13 @@ def to_starlette_response(normalised_response: NormalisedResponse) -> starlette.
     
 
     if complete_body:
-        if (content_type := headers.get("content-type", None)) is not None:
-            return make_response_from_status(status, content_type=content_type)
-    
-    return make_response_from_complete_spec()
-
         
-        
-        
+        content_type = headers.get("content-type", None) # type: ignore
 
+        if content_type == None:
+            return make_response_from_status(status, content_type=content_type, headers=headers)
 
+        return make_response_from_status(status, content_type=content_type)
 
-        
+    return make_response_from_complete_spec(status, headers, body)
 
-
-    
-
-    
-    
-
-    
-
-
